source_directory/inference/requirements.txt
numpy
Pillow
tensorflow-gpu